Updated PostsController class: - Updated CreatePost method in order to not send from the front end the guid for the Attachments attached to a Post -> but this still needs updates for the Mapping and the DTOs sent to the client.
Updated AttachmentDTO class in order to let the DB assing the GUID for the attachment.

Created the location for the Services classes that will be between the Controllers and the Repositories, in the Application layer.

Need to think for a solution for the GetAll methods that at some point will receive a huge ammount of datas.